[Communities] Refactor pallet structure (#312)
* refactor(pallets/communities): resolve ambiguity between CommunityMetadata storage type and struct

* refactor(pallets/communities): split logical modules within pallet

* refactor(communities:pallet): use pallet::call(weight) option

* refactor(pallet/communities): remove excessive verbosity

* change(pallet/communities): update documentation regarding tokens

* refactor(pallets/communities): split types into multiple files
