+++ pallets/communities/README.md
@@ -51,15 +51,11 @@ facilitating its participants to have governance over the community entity
   the community. Can receive [payments][5], transfers, or payment [fees][6].
   It can transfer funds via a privileged call executed by the community
   _admin_ or a call dispatched from a proposal.
-- **Governance Token:** A [non-sufficient fungible asset][7] issued and
-  administered by the _Treasury Account_ of the community. Customarily, it's
-  given among community members and can be used to vote depending on the
-  voting mechanism set by the community.
-- **Economic Token:** A [sufficient fungible asset][7] issued and
-  administered by the _Treasury Account_ of the community. Generally used
-  for monetary purposes, it can be transferred among members and non-members
-  of the community, used to pay network fees, and for [payments][5] and its
-  corresponding [fees][6].
+- **Community Token:** A Community might create and manage tokens by
+  dispatching the respective call from the corresponding origin. These tokens
+  also might be used to vote if set via Voting Mechanism.
+- **Voting Method:** Can be either rank weighed, member-counted, or asset-weighed
+  and determines how the votes of proposals will be tallied.
 
 ## Goals
 
@@ -160,8 +156,9 @@ dispatched through an approved proposal. !
 
 - `community`: Stores the basic information of the community. If a value exists for a
   specified [`ComumunityId`][t00], this means a community exists.
+- `asset_id`: Stores the ID of the [asset][7] for a given [`CommunityId`][t00].
 - `metadata`: Stores the metadata regarding a community.
-- `member_information`: Stores the information of a community (specified by its
+- `membership`: Stores the information of a community (specified by its
   [`CommunityId`][t00]) member (specified by it's [`AccountId`][1]).
 - `members_count`: Store the count of community members. This simplifies the process of keeping track of members' count.
