+++ pallets/communities/src/functions/challenges.rs
@@ -2,8 +2,7 @@ use super::*;
 
 impl<T: Config> Pallet<T> {
 	pub(crate) fn ensure_active(community_id: &CommunityIdOf<T>) -> DispatchResult {
-		let community_info =
-			<CommunityInfo<T>>::try_get(community_id).map_err(|_| Error::<T>::CommunityDoesNotExist)?;
+		let community_info = Self::community(community_id).ok_or(Error::<T>::CommunityDoesNotExist)?;
 
 		if community_info.state != CommunityState::Active {
 			Err(Error::<T>::CommunityNotActive)?
@@ -14,7 +13,7 @@ impl<T: Config> Pallet<T> {
 
 	#[allow(dead_code)]
 	pub(crate) fn do_force_complete_challenge(community_id: &CommunityIdOf<T>) -> DispatchResult {
-		<CommunityInfo<T>>::try_mutate_exists(community_id, |value| {
+		Info::<T>::try_mutate_exists(community_id, |value| {
 			let Some(community_info) = value else {
 				return Err::<(), DispatchError>(Error::<T>::CommunityDoesNotExist.into());
 			};
