+++ pallets/communities/src/functions/registry.rs
@@ -14,9 +14,9 @@ impl<T: Config> Pallet<T> {
 			return Err(Error::<T>::CommunityAlreadyExists.into());
 		}
 
-		<CommunityInfo<T>>::insert(
+		Info::<T>::insert(
 			community_id.clone(),
-			Community {
+			CommunityInfo {
 				admin: who.clone(),
 				state: Default::default(),
 				sufficient_asset_id: None,
@@ -28,11 +28,8 @@ impl<T: Config> Pallet<T> {
 		Ok(())
 	}
 
-	pub(crate) fn do_set_metadata(
-		community_id: &CommunityIdOf<T>,
-		value: types::CommunityMetadata<T>,
-	) -> DispatchResult {
-		<pallet::CommunityMetadata<T>>::try_mutate(community_id, |metadata| {
+	pub(crate) fn do_set_metadata(community_id: &CommunityIdOf<T>, value: CommunityMetadata<T>) -> DispatchResult {
+		Metadata::<T>::try_mutate(community_id, |metadata| {
 			*metadata = Some(value);
 
 			Ok(())
