+++ pallets/communities/src/types/primitives.rs
@@ -1,0 +1,4 @@
+use super::*;
+
+pub type SizedField<S> = BoundedVec<u8, S>;
+pub type ConstSizedField<const S: u32> = SizedField<ConstU32<S>>;
